Stoprail configuration form translation bug correction
The implementation of the translation of the doors opening side in the stoprail
configuration form (commit c08896f) introduced a bug locking the opening doors
side selection dropdown box to "Closed".

This commit corrects this bug by use of the numerical index (1,2,3) of the side
selection instead of its litteral value ("Left","Right","Closed" or their translations).